+++ idea-compat/src/net/groboclown/idea/p4ic/compat/UICompat.java
@@ -17,6 +17,8 @@ package net.groboclown.idea.p4ic.compat;
 import com.intellij.openapi.options.Configurable;
 import com.intellij.openapi.progress.ProgressIndicator;
 import com.intellij.openapi.project.Project;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 public abstract class UICompat {
     public static final UICompat getInstance() {
@@ -38,4 +40,23 @@ public abstract class UICompat {
      * @return true if the configuration was changed
      */
     public abstract boolean editVcsConfiguration(Project project, Configurable configurable);
+
+
+    /**
+     * Ask the user for a password.
+     *
+     * @param project project source
+     * @param title window title
+     * @param message main message to display
+     * @param requester password key owner class
+     * @param key password key
+     * @param resetPassword true if the password should be reset before running; not entering a
+     *                      password will blank it out.
+     * @param error the error message to display
+     * @return the password entered by the user, or null if the user cancelled the action.
+     */
+    @Nullable
+    public abstract String askPassword(@Nullable Project project,
+            @NotNull String title, @NotNull String message, Class<?> requester,
+            @NotNull String key, boolean resetPassword, @NotNull String error);
 }