+++ idea140/src/net/groboclown/idea/p4ic/compat/idea140/UICompat140.java
@@ -15,12 +15,15 @@
 package net.groboclown.idea.p4ic.compat.idea140;
 
 import com.intellij.ide.actions.ShowSettingsUtilImpl;
+import com.intellij.ide.passwordSafe.ui.PasswordSafePromptDialog;
 import com.intellij.openapi.options.Configurable;
 import com.intellij.openapi.options.ShowSettingsUtil;
 import com.intellij.openapi.progress.ProgressIndicator;
 import com.intellij.openapi.progress.ProgressManager;
 import com.intellij.openapi.project.Project;
 import net.groboclown.idea.p4ic.compat.UICompat;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 public class UICompat140 extends UICompat {
     @Override
@@ -35,5 +38,15 @@ public class UICompat140 extends UICompat {
                 ShowSettingsUtilImpl.createDimensionKey(configurable),
                 configurable,
                 true);
+    }
+
+    @Nullable
+    @Override
+    public String askPassword(@Nullable final Project project, @NotNull final String title,
+            @NotNull final String message,
+            final Class<?> requester, @NotNull final String key, final boolean resetPassword,
+            @NotNull final String error) {
+        return PasswordSafePromptDialog.askPassword(project, title, message, requester,
+                key, resetPassword, error);
     }
 }