+++ plugin/src/net/groboclown/idea/p4ic/server/exceptions/PasswordAccessedWrongException.java
@@ -14,28 +14,28 @@
 
 package net.groboclown.idea.p4ic.server.exceptions;
 
-import com.intellij.ide.passwordSafe.PasswordSafeException;
 import com.perforce.p4java.exception.AccessException;
 import com.perforce.p4java.impl.mapbased.rpc.msg.RpcMessage;
 import com.perforce.p4java.impl.mapbased.rpc.msg.ServerMessage;
 import com.perforce.p4java.server.IServerMessage;
 import com.perforce.p4java.server.ISingleServerMessage;
+import net.groboclown.idea.p4ic.P4Bundle;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
 
-public class PasswordStoreException extends AccessException {
-    public PasswordStoreException(@NotNull final PasswordSafeException t) {
-        super(getServiceMessage(t));
-        initCause(t);
+public class PasswordAccessedWrongException extends AccessException {
+    public PasswordAccessedWrongException() {
+        super(getServiceMessage());
     }
 
     @NotNull
-    private static IServerMessage getServiceMessage(final PasswordSafeException t) {
+    private static IServerMessage getServiceMessage() {
         Map<String, Object> format = new HashMap<String, Object>();
-        format.put(RpcMessage.FMT + '0', t.getMessage());
+        format.put(RpcMessage.FMT + '0',
+                P4Bundle.message("error.get-password.in.read.action"));
         return new ServerMessage(Collections.<ISingleServerMessage>singletonList(
                 new ServerMessage.SingleServerMessage("0", 0, format)
             ), format);