+++ plugin/src/net/groboclown/idea/p4ic/v2/server/connection/ConnectionUIConfiguration.java
@@ -14,7 +14,9 @@
 
 package net.groboclown.idea.p4ic.v2.server.connection;
 
+import com.intellij.openapi.project.Project;
 import com.perforce.p4java.exception.P4JavaException;
+import net.groboclown.idea.p4ic.server.exceptions.P4LoginException;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
@@ -30,8 +32,14 @@ import java.util.Map;
  */
 public class ConnectionUIConfiguration {
     public static void checkConnection(@NotNull ProjectConfigSource source)
-            throws P4JavaException, IOException, URISyntaxException {
-        ClientExec.getServerInfo(source.getServerConfig());
+            throws P4JavaException, IOException, URISyntaxException, P4LoginException {
+        try {
+            ClientExec.getServerInfo(source.getProject(), source.getServerConfig());
+        } catch (P4LoginException e) {
+            ClientExec.loginFailure(source.getProject(), source.getServerConfig(),
+                    NOOP_CONTROLLER, e.getP4JavaException());
+            throw e;
+        }
     }
 
     @SuppressWarnings("ThrowableResultOfMethodCallIgnored")
@@ -48,6 +56,10 @@ public class ConnectionUIConfiguration {
                 ret.put(source, e);
             } catch (URISyntaxException e) {
                 ret.put(source, e);
+            } catch (P4LoginException e) {
+                ClientExec.loginFailure(source.getProject(), source.getServerConfig(),
+                        NOOP_CONTROLLER, e.getP4JavaException());
+                ret.put(source, e);
             }
         }
         return ret;
@@ -62,13 +74,17 @@ public class ConnectionUIConfiguration {
         }
         for (ProjectConfigSource source : sources) {
             try {
-                final List<String> clients = ClientExec.getClientNames(source.getServerConfig());
+                final List<String> clients = ClientExec.getClientNames(source.getProject(), source.getServerConfig());
                 ret.put(source, new ClientResult(clients));
             } catch (IOException e) {
                 ret.put(source, new ClientResult(e));
             } catch (P4JavaException e) {
                 ret.put(source, new ClientResult(e));
             } catch (URISyntaxException e) {
+                ret.put(source, new ClientResult(e));
+            } catch (P4LoginException e) {
+                ClientExec.loginFailure(source.getProject(), source.getServerConfig(),
+                        NOOP_CONTROLLER, e.getP4JavaException());
                 ret.put(source, new ClientResult(e));
             }
         }
@@ -101,6 +117,42 @@ public class ConnectionUIConfiguration {
 
         public Exception getConnectionProblem() {
             return connectionProblem;
+        }
+    }
+
+
+    private static final NoOpServerConnectedController NOOP_CONTROLLER = new NoOpServerConnectedController();
+
+    private static class NoOpServerConnectedController implements ServerConnectedController {
+
+        @Override
+        public boolean isWorkingOnline() {
+            return false;
+        }
+
+        @Override
+        public boolean isWorkingOffline() {
+            return false;
+        }
+
+        @Override
+        public boolean isAutoOffline() {
+            return false;
+        }
+
+        @Override
+        public boolean isValid() {
+            return false;
+        }
+
+        @Override
+        public void disconnect() {
+
+        }
+
+        @Override
+        public void connect(@NotNull final Project project) {
+
         }
     }
 }